fix: corrige errores críticos en evaluación GPU y operadores
Se corrigen cuatro bugs en el sistema de evaluación GPU:

- FIX N1: La función gamma ahora calcula correctamente exp(lgamma(a)) en lugar de lgamma(a+1), que computaba log-factorial en vez de Gamma.
- FIX N2: Se implementa el operador lgamma faltante; antes las fórmulas con lgamma evaluaban silenciosamente a 0.
- FIX N3: Se añade protección contra transposición incorrecta de matrices cuadradas cuando num_vars == num_samples.
- FIX N4: Se corrige overflow silencioso en el cálculo de hashes al convertir population a long() antes de multiplicar por pesos int64.

+++ AlphaSymbolic/core/gpu/operators.py
@@ -725,7 +725,10 @@ class GPUOperators:
         else:
              weights = self.dedup_weights
              
-        hashes = (population * weights).sum(dim=1)
+        # FIX N4: population is uint8; multiplying by int64 weights in uint8 causes
+        # silent overflow (e.g. 200*256 mod 256 = 0), producing false hash collisions.
+        # Cast to long() first so the full int64 range is used for hashing.
+        hashes = (population.long() * weights).sum(dim=1)
         
         # 2. Unique on Hashes (1D is fast)
         # return_inverse gives indices such that hashes = unique[inverse]
